requirements.txt
pytoch == 1.12.1
pytorch_lighting == 1.7
numpy == 1.23.5
matplotlib == 3.6.2
pandas == 1.5.2